requirements.txt
asgiref==3.6.0
autopep8==2.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.0
defusedxml==0.7.1
dj-database-url==1.2.0
Django==3.2
django-cors-headers==3.7.0
django-environ==0.7.0
django-filter==21.1
django-summernote==0.8.11.6
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
gunicorn==20.1.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
oauthlib==3.2.2
Pillow==8.3.2
psycopg2-binary==2.9.1
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.1.0
python-http-client==3.3.7
python3-openid==3.2.0
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
sendgrid==6.9.7
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
starkbank-ecdsa==2.2.0
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.14
whitenoise==6.4.0
